Revert "#350 Improve GeD transactions."
This reverts commit a05b192522712099a454d735c975f18bd88fd516.

+++ web/modules/custom/girchi_ged_transactions/src/Entity/GedTransaction.php
@@ -217,7 +217,7 @@ class GedTransaction extends ContentEntityBase implements GedTransactionInterfac
       ->setDescription(t('The time that the entity was last edited.'));
 
     // My custom fields.
-    $fields['ged_amount'] = BaseFieldDefinition::create('integer')
+    $fields['ged_amount'] = BaseFieldDefinition::create('decimal')
       ->setLabel(t('Ged amount'))
       ->setTranslatable(TRUE)
       ->setDisplayOptions('form', [
